requirements.txt
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
numpy==1.26.4
matplotlib==3.8.4
opencv-python==4.9.0.80
vosk==0.3.44
sounddevice==0.4.6
qiskit==1.0.2
pyttsx3==2.90
Flask==2.3.3


